Began work on a Jodin Interpreter. Reworked bootstrap plan from Java Compiler -> Jodin Compiler to: Java Interpreter -> Jodin Compiler
This will allow for easier creation of the Jodin compiler.
The way this will be done is by first programming the interpreter in Java, then writing the compiler in Jodin, running the compiler in the java-based interpreter, then compiling the compiler within the compiler running in the java interpreter.
This process will produce a compiler that is self hosting.