Continued work into interpreter and how it runs
It uses a stack based memory of sorts, with a separate return value register

Initialization will be handled using java. Because malloc and calloc are therefore useless in the
Interpreter, it is wise to remove them from prelude. Furthermore, since void* to type* casts are
disabled, the presence of these functions are redundant.

As such, the stack and the heap are essentially managed by Java.
For classes, all methods will be dynamically found.

Signed-off-by: FreedomFries37 <30393678+FreedomFries37@users.noreply.github.com>